Fix build on macOS
The compiled Python module of pyexiv2 is dynamically linked to the
libexiv2.27.dylib and bundles this library as libexiv2.dylib.

pyexiv2 loads this dylib dynamically in its __init__.py, but this seems
to stop working when packaging iscc-cli with PyInstaller.

Based on help from the PyInstaller maintainers, the build script now
does some fun hacks in the pyexiv2 site-packages to clean this up and
produce a working executable.

See: https://github.com/pyinstaller/pyinstaller/discussions/6404